requirements.txt
fastapi==0.109.2
pydantic==1.9.0
aiohttp
sqlalchemy==2.0.27
gunicorn==21.2.0
uvicorn==0.27.1
python-dotenv==1.0.1
aioredis==1.3.1
alembic==1.7.7
psycopg2-binary==2.9.3
